Completed Assignment 1 of Week 1
Implemented code for following Assignment

-- Task 1 --
1. Complete writing the method getNumPoints that has one parameter s that is of type Shape.
This method  returns an integer that is the number of points in Shape s.

2. Add code in the method testPerimeter to call getNumPoints and to print the result.

-- Task 2 --
3. Complete writing the method getAverageLength that has one parameter s that is of type Shape.
This method returns a number of type double that is the calculated average of all the sides’ lengths in the Shape S.

4. Add code in the method testPerimeter to call the method getAverageLength and to print out the result.
Note if you were to select the file example1.txt, then the average side length should be 4.0.

-- Task 3 --
5. Complete writing the method getLargestSide that has one parameter s that is of type Shape.
This method returns a number of type double that is the longest side in the Shape S.

6. Add code in the method testPerimeter to call the method getLargestSide and to print out the result.
Note if you were to select the file example1.txt, then the longest side should be 5.0.

-- Task 4 --
7. Complete writing the method getLargestX that has one parameter s that is of type Shape.
This method returns a number of type double that is the largest x value over all the points in the Shape s.

8. Add code in the method testPerimeter to call the method getLargestX and to print out the result.
Note if you were to select the file example1.txt, then the longest side should be 4.0.