completed Assignment 2 of week 1
Implemented code for following Assignment

-- Task 1 --
1. Complete writing the method getLargestPerimeterMultipleFiles that has no parameters.
This method creates a DirectoryResource (so you can select multiple files) and
then iterates over these files. For each File f, it converts the file into a FileResource with the line
FileResource fr = new FileResource(f);
Then it creates a Shape from the FileResource and calculates the shapes perimeter.
What else does it need to do? It needs to return the the largest perimeter over all the shapes
in the files you have selected.

2. Finish writing the void method testPerimeterMultipleFiles to call getLargestPerimeterMultipleFiles
and to print out the largest such perimeter.
This method has no parameters and no return value.
You will select the files when you run this method
(hint: see our documentation for the DirectoryResource class).

-- Task 2 --

3. Finish writing the method getFileWithLargestPerimeter that has no parameters.
This method should, like the getLargestPerimeterMultipleFiles method,
create its own Directory Resource, except that this new method returns the File
that has the largest such perimeter, so it has return type File.

4. Add code to the method testFileWithLargestPerimeter to call getFileWithLargestPerimeter.
For the File that is returned, print the name of that file.